Add missing UCI protocol definitions and fix project functionality
- Create missing uci_pdl.h header with UCI protocol constants
- Add missing handle_core_device_status_ntf function implementation
- Fix buffer overflow issues in send_uci_command
- Update README to accurately describe UWB UCI protocol interface
- Add function declaration type fixes
- Create test script to demonstrate functionality

+++ src/uci_functions.h
@@ -3,7 +3,9 @@
 
 #include "uci.h"
 
+#include <stddef.h>
+
 void send_uci_command(unsigned char mt, unsigned char pbf, unsigned char gid, unsigned char oid, unsigned char* payload, int payload_len);
-void parse_uci_packet(unsigned char* packet, int packet_len);
+void parse_uci_packet(unsigned char* packet, size_t packet_len);
 
 #endif // UCI_FUNCTIONS_H
